Use the git archive to export the files
Previously the maketar.sh just did a recursive copy to make the tarball.
This version uses the git status to make sure the repositories have everything
checked in before attempting to make the tarball. The "git archive" is used
to copy the files over. This ensures that the timestamps are preserved.